typescriptの変更に伴って出たエラーを解消中
