+++ src/Category/Category.tsx
@@ -1,5 +1,5 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
+import * as React from 'react';
+// import { connect } from 'react-redux';
 import axios from "axios";
 // import './Category.scss';
 import {Table, TableBody, TableHead, TableCell, TableRow} from '@material-ui/core';
@@ -10,8 +10,16 @@ import EditModal from './EditModal';
 import DelModal from './DelModal';
 import Item from '../Item/Item';
 
-class Category extends Component {
-    constructor(props) {
+interface categoryProps {
+    categories: string,
+    Component: string,
+    isCategoryAddModal: boolean,
+    isCategoryEditModal: boolean,
+    isCategoryDelModal: boolean
+}
+
+export default class Category extends React.Component<categoryProps> {
+    constructor(props: categoryProps) {
         super(props);
         this.state = {
             categories: [],
@@ -27,6 +35,7 @@ class Category extends Component {
     }
     isShowEditModal(boolean, Id) {
         this.setState({ isCategoryEditModal: boolean, categoryId: Id })
+        // console.log('カテゴリー.jsのbool：' + this.state.isCategoryEditModal)
     }
     isShowDelModal(boolean) {
         this.setState({ isCategoryDelModal: boolean })
@@ -36,13 +45,13 @@ class Category extends Component {
     componentDidMount() {
         axios.get("/api/categories").then((res) => {
             this.setState({ categories: res.data })
-            console.log(this.state.categories)
+            // console.log(this.state.categories)
         });
     }
 
     render() {
-        const {Component} = this.state;
-        if(Component) return <Component />;
+        // const {Component} = this.state;
+        // if(Component) return <Component />;
         return (
             <Table>
                 <TableHead>
@@ -64,7 +73,8 @@ class Category extends Component {
                     <AddModal />
                 )}
                 { this.state.isCategoryEditModal && (
-                    <EditModal　/>
+                    // <EditModal onClose={() => this.isShowEditModal(false)} />
+                    <EditModal isOpen={this.state.isCategoryEditModal} onClose={this.isShowEditModal(false)} />
                 )}
                 { this.state.isCategoryDelModal && (
                     <DelModal />
@@ -73,5 +83,3 @@ class Category extends Component {
         )
     }
 }
-
-export default connect((state) => state)(Category);
